+++ tests/test_from_thread.py
@@ -6,7 +6,7 @@ import threading
 import time
 from collections.abc import Awaitable, Callable
 from concurrent import futures
-from concurrent.futures import CancelledError
+from concurrent.futures import CancelledError, Future
 from contextlib import asynccontextmanager, suppress
 from contextvars import ContextVar
 from typing import Any, AsyncGenerator, Literal, NoReturn, TypeVar
@@ -63,6 +63,52 @@ def thread_worker_async(
 def thread_worker_sync(func: Callable[..., T_Retval], *args: Any) -> T_Retval:
     assert threading.current_thread() is not threading.main_thread()
     return from_thread.run_sync(func, *args)
+
+
+async def test_thread_cancelled() -> None:
+    event = threading.Event()
+    thread_finished_future: Future[None] = Future()
+
+    def sync_function() -> None:
+        event.wait(3)
+        try:
+            from_thread.check_cancelled()
+        except BaseException as exc:
+            thread_finished_future.set_exception(exc)
+        else:
+            thread_finished_future.set_result(None)
+
+    async with create_task_group() as tg:
+        tg.start_soon(to_thread.run_sync, sync_function)
+        await wait_all_tasks_blocked()
+        tg.cancel_scope.cancel()
+        event.set()
+
+    with pytest.raises(get_cancelled_exc_class()):
+        thread_finished_future.result(3)
+
+
+async def test_thread_cancelled_and_abandoned() -> None:
+    event = threading.Event()
+    thread_finished_future: Future[None] = Future()
+
+    def sync_function() -> None:
+        event.wait(3)
+        try:
+            from_thread.check_cancelled()
+        except BaseException as exc:
+            thread_finished_future.set_exception(exc)
+        else:
+            thread_finished_future.set_result(None)
+
+    async with create_task_group() as tg:
+        tg.start_soon(lambda: to_thread.run_sync(sync_function, abandon_on_cancel=True))
+        await wait_all_tasks_blocked()
+        tg.cancel_scope.cancel()
+
+    event.set()
+    with pytest.raises(get_cancelled_exc_class()):
+        thread_finished_future.result(3)
 
 
 class TestRunAsyncFromThread: